requirements.txt
web3
flare-python-periphery-package # TODO
python-dotenv
